Apply patch to fix CVE-2024-53858 (Closes: #1088808)
Git credentials helpers are now configured to only be enabled for the
matching host that should be supplied the authentication tokens to. This
will prevent gh from leaking tokens for GitHub and GitHub Enterprise
Server to other hosts when recursively cloning submodules in a
repository. This patch is based on the upstream commits from 46922694 to
c94def8b, inclusively.